// src/app/models/KKC/kkc-dbaShareReleaseNote.ts
export class kkcdbaEntry {
    unique: string; //unique guid
    dbaId: string; // automatic 
    dbaNo: string;
    dbaDate: string;
    centerId: string;
    centerCode: string;
    courseCode: string;
    feesItem: string;
    despatchNo: string;
    batchNo: string;
    despatchMonth: string;
    despatchDate: string;
    despatchAmount: string;
    tax: number;
    feeWithoutTax: number;
    stkAmount: string; // share to kkc
    stkRate: string;
    enteredBy: string;
    isdbaEntered: boolean;


}